Last Updates

+++ app/Http/Controllers/Api/V1/Documents/Documents.php
@@ -9,6 +9,7 @@ use App\Traits\Api\V1\Responses;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Spatie\Activitylog\Models\Activity;
 
 class Documents extends Controller
 {
@@ -20,7 +21,42 @@ class Documents extends Controller
     {
         $docs = ModelsDocuments::with('user')->paginate(10);
 
-        return $this->success(status: Response::HTTP_OK, message: 'Document Retrieved Successfully.', data: $docs);
+        return $this->successPaginated(status: Response::HTTP_OK, message: 'Document Retrieved Successfully.', data: $docs);
+    }
+
+    public function getAllDataWithoutPaginate(){
+        $data = ModelsDocuments::with('user')->get();
+
+        return $this->success(status: Response::HTTP_OK, message: 'Currency Retrieved Successfully.', data: $data);
+    }
+
+    public function getLogs(string $id){
+        $data = ModelsDocuments::find($id);
+
+        if(!$data){
+            return $this->error(
+                status: Response::HTTP_NOT_FOUND,
+                message: "Sorry, the requested data was not found."
+            );
+        }
+
+        $logs = Activity::where('subject_id', $data->id)
+                        ->where('subject_type', ModelsDocuments::class)
+                        ->get();
+
+        if ($logs->isEmpty()) {
+            return $this->error(
+                status: Response::HTTP_NOT_FOUND,
+                message: "No logs found for the specified Document."
+            );
+        }
+
+
+        return $this->success(
+            status:Response::HTTP_OK
+            , message:'Logs Retrived Succesfuly'
+            , data:  $logs
+        );
     }
 
     /**
@@ -51,7 +87,7 @@ class Documents extends Controller
      */
     public function show(string $id)
     {
-        $docs = ModelsDocuments::with('user' )->where('id' , $id)->first();
+        $docs = ModelsDocuments::with('user')->where('id' , $id)->first();
 
         if (!$docs) {
             return $this->error(status: Response::HTTP_INTERNAL_SERVER_ERROR, message: 'docs not found.',);
@@ -73,8 +109,12 @@ class Documents extends Controller
             return $this->error(status: Response::HTTP_INTERNAL_SERVER_ERROR, message: 'docs not found.',);
         }
 
+        if($docs->user_id != auth()->guard('api')->id()){
+            return $this->error(status: Response::HTTP_INTERNAL_SERVER_ERROR, message: 'Only User Who Created The Docs Can Update It.',);
+        }
+
         $docs->update($request->except('attachments' , 'demo_file' , 'final_file'));
-        
+
         $docs->clearMediaCollection('docs');
         $docs->addMediaFromRequest('attachments')->toMediaCollection('docs', 'docs');
         $docs->addMediaFromRequest('demo_file')->toMediaCollection('docs', 'docs');
@@ -95,6 +135,11 @@ class Documents extends Controller
         if (!$docs) {
             return $this->error(status: Response::HTTP_INTERNAL_SERVER_ERROR, message: 'docs not found.',);
         }
+
+        if($docs->user_id != auth()->guard('api')->id()){
+            return $this->error(status: Response::HTTP_INTERNAL_SERVER_ERROR, message: 'Only User Who Created The Docs Can Update It.',);
+        }
+
         $docs->clearMediaCollection('docs');
 
         $docs->delete();
@@ -102,5 +147,44 @@ class Documents extends Controller
         return $this->success(status: Response::HTTP_OK, message: 'Document Deleted Successfully.', data: [
             'docs' => $docs,
         ]);
+    }
+
+
+    public function restore(string $id)
+    {
+        $data = ModelsDocuments::withTrashed()->find($id);
+
+        if (!$data) {
+            return $this->error(
+                status: Response::HTTP_NOT_FOUND,
+                message: 'Document not found.'
+            );
+        }
+
+        if (!$data->trashed()) {
+
+            return $this->error(
+                status: Response::HTTP_BAD_REQUEST,
+                message: 'Document not found.'
+            );
+        }
+
+        $data->restore();
+
+        return $this->success(
+            status:Response::HTTP_OK
+            ,message: 'Document restored successfully.'
+            , data: $data
+        );
+    }
+
+    public function getAllTrashedData(){
+        $data = ModelsDocuments::onlyTrashed()->paginate(10);
+
+        return $this->successPaginated(
+            status:Response::HTTP_OK
+            , message:'Document Retrived Succesfuly'
+            , data: $data
+        );
     }
 }
